[Task 6 - Selenium Grid]
1. Test scripts are run from the TestBddRunner class

2. From the 'getFeaturesTags' method
we can get tags that determine which test scripts to run
(these are scripts with the same tags).

These tags can be taken from any data source,
for example, from the 'properties' file or 'excel table'

3. In the annotation @CucumberOptions
the paths to the directories where all the classes 'fetures' and 'steps' are located.
We can also specify the tags parameter there
(test scripts marked with these tags will run)

4. In the testng.xml file, it is possible to specify the configurations of test classes that can be run for execution
(in this project, such a test class is only 'TestBddRunner')

Thus, the launch of test scripts can be adjusted as follows:

1. Using the @CucumberOptions annotation
(specify paths to 'features' and 'steps' files, as well as tags in the 'tags' parameter)

2. By creating several runners classes,
which will run test cases grouped logically,
for example, only tests to check 'login' and 'registration'

Such test scripts can be run
using configurations from testng.xml

3. In the runner class, by passing tags to the 'scenario' method
and these tags determine which test scripts to run

This approach is implemented in this project:
test scripts are launched with @LoginUserProfile and @TabletsSearch tags
in the runner class 'TestBddRunner'